Add present value calculation method

+++ java-playground-aci/src/main/java/com/andymur/pg/java/aci/BasePeriod.java
@@ -11,7 +11,7 @@ public enum BasePeriod {
 		this.numberOfDays = numberOfDays;
 	}
 
-	public int getNumberOfDays() {
+	public long getNumberOfDays() {
 		return numberOfDays;
 	}
 }
